extracted DataInfo class to its own file
