+++ src/app/data-loader/workspace-info.ts
@@ -1,3 +1,5 @@
+import { DataInfo } from './DataInfo';
+
 import {zip, invert} from '../util/util'
 import { Synchronizer } from '../util/sync';
 
@@ -118,53 +120,6 @@ export class VideoInfo {
 
     // #region [Public Methods]
     sync(data: DataInfo): Synchronizer { return new Synchronizer(data.flashes, this.flashes) }
-    // #endregion
-}
-
-export class DataInfo {
-    // #region [Constants]
-    static MILLISECONDS = 1000
-    // #endregion
-
-    // #region [Properties]
-    workspace: string;
-    name: string;
-    format: string;
-    labelled: boolean | string;
-    Hz: number;
-    channels: string;
-    path: string;
-    flashes: number[];
-    hide: boolean;
-    type: string;
-    crop?: number[];
-    private ws: WorkspaceInfo;
-    // #endregion
-
-    // #region [Constructor]
-    constructor(ws, info) {
-        // provided properties
-        this.ws = ws;
-        this.workspace = info.workspace;
-        this.name = info.name;
-        // required properties
-        this.format = info.format;
-        this.labelled = info.labelled;
-        this.Hz = info.Hz;
-        this.channels = info.channels;
-        this.path = info.path;
-        // optional properties
-        this.flashes = info.flashes || [];
-        this.hide = info.hide || false;
-        this.type = info.type || 'data';
-        if ('crop' in info) this.crop = info.crop;
-    }
-    // #endregion
-
-    // #region [Accessors]
-    get rate() { return DataInfo.MILLISECONDS / this.Hz }
-
-    get isEnergy() { return this.type === 'energy' }
     // #endregion
 }
