+++ src/app/dataview/dataview.component.ts
@@ -3,7 +3,8 @@ import { ActivatedRoute } from '@angular/router';
 import { Component, OnInit, HostListener, ViewChildren, ViewChild, 
          ChangeDetectorRef, AfterViewChecked } from '@angular/core';
 import { DataloaderService } from '../data-loader/data-loader.service';
-import { WorkspaceInfo, DataInfo, TypeMap, LabelScheme } from '../data-loader/workspace-info';
+import { WorkspaceInfo, TypeMap, LabelScheme } from '../data-loader/workspace-info';
+import { DataInfo } from "../data-loader/DataInfo";
 import { SettingsService } from '../settings/settings.service';
 import { Label, LabelStream} from './labelstreams/labelstream';
 import { ModeTracker } from './modes/tool-mode';
