+++ src/app/dataview/energy/popover/compute-popover.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
-import { WorkspaceInfo, DataInfo } from '../../../data-loader/workspace-info';
+import { WorkspaceInfo } from '../../../data-loader/workspace-info';
+import { DataInfo } from "../../../data-loader/DataInfo";
 import { SaverService } from '../../../saver/saver.service';
 
 @Component({
