+++ src/app/dataview/video/video.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Input, ElementRef, ChangeDetectorRef, AfterViewChecked, HostListener, Output, EventEmitter } from '@angular/core';
-import { WorkspaceInfo, DataInfo, VideoInfo } from '../../data-loader/workspace-info';
+import { WorkspaceInfo, VideoInfo } from '../../data-loader/workspace-info';
+import { DataInfo } from "../../data-loader/DataInfo";
 import { VgAPI } from 'videogular2/core';
 import { Synchronizer } from '../../util/sync';
 import { zip } from '../../util/util';
