+++ src/app/home/workspace/workspace.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { WorkspaceInfo, DataInfo, LabelScheme } from '../../data-loader/workspace-info';
+import { WorkspaceInfo, LabelScheme } from '../../data-loader/workspace-info';
+import { DataInfo } from "../../data-loader/DataInfo";
 
 // #region [Interfaces]
 interface SelectedInfo {
